+++ romancal/regtest/test_wfi_dq_init.py
@@ -1,3 +1,4 @@
+"""Tests for the DQ Init module and DMS 25 and DMS 26 requirements"""
 import os
 import pytest
 
@@ -12,19 +13,23 @@ def test_dq_init_image_step(rtdata, ignore_asdf_paths):
     # DMS25 Test: Testing retrieval of best ref file for image data,
     # and creation of a ramp file with CRDS selected mask file applied.
 
-    rtdata.get_data("WFI/image/r0000101001001001001_01101_0001_WFI01_uncal.asdf")
-    rtdata.input = "r0000101001001001001_01101_0001_WFI01_uncal.asdf"
+    input_file = "r0000101001001001001_01101_0001_WFI01_uncal.asdf"
+    rtdata.get_data(f"WFI/image/{input_file}")
+    rtdata.input = input_file
 
     # Test CRDS
     step = DQInitStep()
     model = rdm.open(rtdata.input)
-    step.log.info('DMS25 MSG: Testing retrieval of best ref file for image data, '
+    step.log.info('DMS25 MSG: Testing retrieval of best '
+                  'ref file for image data, '
                   'Success is creation of a ramp file with CRDS selected '
                   'mask file applied.')
 
-    step.log.info(f'DMS25 MSG: First data file: {rtdata.input.rsplit("/", 1)[1]}')
+    step.log.info('DMS25 MSG: First data file: '
+                  f'{rtdata.input.rsplit("/", 1)[1]}')
     ref_file_path = step.get_reference_file(model, "mask")
-    step.log.info(f'DMS25 MSG: CRDS matched mask file: {ref_file_path.rsplit("/", 1)[1]}')
+    step.log.info('DMS25 MSG: CRDS matched mask file: '
+                  f'{ref_file_path.rsplit("/", 1)[1]}')
     ref_file_name = os.path.split(ref_file_path)[-1]
 
     assert "roman_wfi_mask" in ref_file_name
@@ -38,15 +43,17 @@ def test_dq_init_image_step(rtdata, ignore_asdf_paths):
                   ' not having been implemented yet.')
     RomanStep.from_cmdline(args)
     ramp_out = rdm.open(rtdata.output)
-    step.log.info(f'DMS25 MSG: Does ramp data contain pixeldq from mask file? : '
+    step.log.info(f'DMS25 MSG: Does ramp data contain '
+                  'pixeldq from mask file? : '
                   f'{("roman.pixeldq" in ramp_out.to_flat_dict())}')
     assert "roman.pixeldq" in ramp_out.to_flat_dict()
 
     rtdata.get_truth(f"truth/WFI/image/{output}")
-    step.log.info(f'DMS25 MSG: Was the proper data quality array initialized'
+    step.log.info('DMS25 MSG: Was the proper data quality array initialized'
                   ' for the ramp data produced? : '
                   f'{(compare_asdf(rtdata.output, rtdata.truth, **ignore_asdf_paths) is None)}')
-    assert compare_asdf(rtdata.output, rtdata.truth, **ignore_asdf_paths) is None
+    assert compare_asdf(rtdata.output, rtdata.truth,
+                        **ignore_asdf_paths) is None
 
 
 @pytest.mark.bigdata
@@ -54,18 +61,23 @@ def test_dq_init_grism_step(rtdata, ignore_asdf_paths):
     """DMS26 Test: Testing retrieval of best ref file for grism data,
      and creation of a ramp file with CRDS selected mask file applied."""
 
-    rtdata.get_data("WFI/grism/r0000101001001001001_01102_0001_WFI01_uncal.asdf")
-    rtdata.input = "r0000101001001001001_01102_0001_WFI01_uncal.asdf"
+    input_file = "r0000101001001001001_01102_0001_WFI01_uncal.asdf"
+    rtdata.get_data(f"WFI/grism/[input_file]")
+    rtdata.input = input_file
 
     # Test CRDS
     step = DQInitStep()
     model = rdm.open(rtdata.input)
-    step.log.info('DMS26 MSG: Testing retrieval of best ref file for grism data, '
-                  'Success is creation of a ramp file with CRDS selected mask file applied.')
+    step.log.info('DMS26 MSG: Testing retrieval of best '
+                  'ref file for grism data, '
+                  'Success is creation of a ramp file with CRDS selected '
+                  'mask file applied.')
 
-    step.log.info(f'DMS26 MSG: First data file: {rtdata.input.rsplit("/", 1)[1]}')
+    step.log.info(f'DMS26 MSG: First data file: '
+                  f'{rtdata.input.rsplit("/", 1)[1]}')
     ref_file_path = step.get_reference_file(model, "mask")
-    step.log.info(f'DMS26 MSG: CRDS matched mask file: {ref_file_path.rsplit("/", 1)[1]}')
+    step.log.info(f'DMS26 MSG: CRDS matched mask file: '
+                  f'{ref_file_path.rsplit("/", 1)[1]}')
     ref_file_name = os.path.split(ref_file_path)[-1]
 
     assert "roman_wfi_mask" in ref_file_name
@@ -79,11 +91,14 @@ def test_dq_init_grism_step(rtdata, ignore_asdf_paths):
                   'not having been implemented yet.')
     RomanStep.from_cmdline(args)
     ramp_out = rdm.open(rtdata.output)
-    step.log.info(f'DMS26 MSG: Does ramp data contain pixeldq from mask file? : '
+    step.log.info('DMS26 MSG: Does ramp data contain pixeldq '
+                  'from mask file? : '
                   f'{("roman.pixeldq" in ramp_out.to_flat_dict())}')
     assert "roman.pixeldq" in ramp_out.to_flat_dict()
 
     rtdata.get_truth(f"truth/WFI/grism/{output}")
-    step.log.info(f'DMS26 MSG: Was proper data quality initialized ramp data produced? : '
+    step.log.info('DMS26 MSG: Was proper data quality initialized '
+                  'ramp data produced? : '
                   f'{(compare_asdf(rtdata.output, rtdata.truth, **ignore_asdf_paths) is None)}')
-    assert compare_asdf(rtdata.output, rtdata.truth, **ignore_asdf_paths) is None
+    assert compare_asdf(rtdata.output, rtdata.truth,
+                        **ignore_asdf_paths) is None