+++ romancal/regtest/test_wfi_flat_field.py
@@ -9,6 +9,7 @@ from .regtestdata import compare_asdf
 
 @pytest.mark.bigdata
 def test_flat_field_image_step(rtdata, ignore_asdf_paths):
+    """Test for the flat field step using imaging data."""
 
     input_data = "r0000101001001001001_01101_0001_WFI01_rampfitstep.asdf"
     rtdata.get_data(f"WFI/image/{input_data}")
@@ -26,12 +27,14 @@ def test_flat_field_image_step(rtdata, ignore_asdf_paths):
     args = ["romancal.step.FlatFieldStep", rtdata.input]
     RomanStep.from_cmdline(args)
     rtdata.get_truth(f"truth/WFI/image/{output}")
-    assert compare_asdf(rtdata.output, rtdata.truth, **ignore_asdf_paths) is None
+    assert compare_asdf(rtdata.output, rtdata.truth,
+                        **ignore_asdf_paths) is None
 
 
 @pytest.mark.skip(reason="There are no grism flats.")
 @pytest.mark.bigdata
 def test_flat_field_grism_step(rtdata, ignore_asdf_paths):
+    """Test for the flat field step using grism data."""
 
     rtdata.get_data("WFI/grism/l2_0005_grism_rate.asdf")
     rtdata.input = "l2_0005_grism_rate.asdf"
@@ -49,14 +52,15 @@ def test_flat_field_grism_step(rtdata, ignore_asdf_paths):
     args = ["romancal.step.FlatFieldStep", rtdata.input]
     RomanStep.from_cmdline(args)
     rtdata.get_truth(f"truth/WFI/grism/{output}")
-    assert (compare_asdf(rtdata.output, rtdata.truth, **ignore_asdf_paths) is None)
+    assert (compare_asdf(rtdata.output, rtdata.truth,
+                         **ignore_asdf_paths) is None)
 
 
 @pytest.mark.bigdata
 @pytest.mark.soctests
 def test_flat_field_crds_match_image_step(rtdata, ignore_asdf_paths):
-    # DMS79 Test: Testing that different datetimes pull different
-    # flat files and successfully make level 2 output
+    """DMS79 Test: Testing that different datetimes pull different
+       flat files and successfully make level 2 output"""
 
     # First file
     input_l2_file = "r0000101001001001001_01101_0001_WFI01_assign_wcs.asdf"
@@ -92,9 +96,11 @@ def test_flat_field_crds_match_image_step(rtdata, ignore_asdf_paths):
     RomanStep.from_cmdline(args)
     rtdata.get_truth(f"truth/WFI/image/{output}")
 
-    step.log.info(f'DMS79 MSG: Was proper flat fielded Level 2 data produced? : '
+    step.log.info('DMS79 MSG: Was proper flat fielded '
+                  'Level 2 data produced? : '
                   f'{(compare_asdf(rtdata.output, rtdata.truth, **ignore_asdf_paths) is None)}')
-    assert compare_asdf(rtdata.output, rtdata.truth, **ignore_asdf_paths) is None
+    assert compare_asdf(rtdata.output, rtdata.truth,
+                        **ignore_asdf_paths) is None
 
     # This test requires a second file, in order to meet the DMS79 requirement.
     # The test will show that two files with different observation dates match
@@ -131,13 +137,16 @@ def test_flat_field_crds_match_image_step(rtdata, ignore_asdf_paths):
                   'implemented yet.')
     RomanStep.from_cmdline(args)
     rtdata.get_truth(f"truth/WFI/image/{output}")
-    step.log.info(f'DMS79 MSG: Was proper flat fielded Level 2 data produced? : '
+    step.log.info('DMS79 MSG: Was proper flat fielded '
+                  'Level 2 data produced? : '
                   f'{(compare_asdf(rtdata.output, rtdata.truth, **ignore_asdf_paths) is None)}')
-    assert compare_asdf(rtdata.output, rtdata.truth, **ignore_asdf_paths) is None
+    assert compare_asdf(rtdata.output, rtdata.truth,
+                        **ignore_asdf_paths) is None
 
     # Test differing flat matches
-    step.log.info(f'DMS79 MSG REQUIRED TEST: Are the two data files matched to different '
-                  f'flat files? : '
-                  f'{("/".join(ref_file_path.rsplit("/", 3)[1:])) != ("/".join(ref_file_path_b.rsplit("/", 3)[1:]))}')
+    step.log.info('DMS79 MSG REQUIRED TEST: Are the two data files '
+                  'matched to different flat files? : '
+                  f'{("/".join(ref_file_path.rsplit("/", 3)[1:]))} != '
+                  f'{("/".join(ref_file_path_b.rsplit("/", 3)[1:]))}')
     assert ("/".join(ref_file_path.rsplit("/", 1)[1:])) != \
            ("/".join(ref_file_path_b.rsplit("/", 1)[1:]))